Universal Two Motor Robot Code
Not fully tested, but should work perfectly fine. It is essentially a
shell of a robot program with autonomous selectors, datadisplay, and a
two motor Chassis. It is all that we are able to do while everybody else
is prototyping. We have nothing else to code until a new mechanical
system is decided upon or finished.