MANY ISSUES HAVE EVOLVED
For some reason, we have run into huge issues revolving around the gyro,
autonomous and other. We managed to fix the gyro which was due to a
WPIlib change this year. However, the sendable chooser remains
mysteriously broken, and two previously absent errors have evolved in
the final line of robot.cpp and structural code of commandbase.cpp.
Those two errors are enormously confusing as they have not had any
issues for the entire season.